refactor: remove temp type and define it as a new type in package
