+++ packages/vscode/src/types/Commit.ts
@@ -1,0 +1,16 @@
+import type { DiffStatistics } from "./DiffStatistics";
+import type { GitHubUser } from "./GithubUser";
+
+export type Commit = {
+  id: string;
+  parentIds: string[];
+  author: GitHubUser;
+  committer: GitHubUser;
+  authorDate: string;
+  commitDate: string;
+  diffStatistics: DiffStatistics;
+  message: string;
+  tags: string[];
+  releaseTags: string[];
+  // fill necessary properties...
+};
