+++ packages/vscode/src/types/DiffStatistics.ts
@@ -1,0 +1,11 @@
+export type DiffStatistics = {
+  changedFileCount: number;
+  insertions: number;
+  deletions: number;
+  files: {
+    [id: string]: {
+      insertions: number;
+      deletions: number;
+    };
+  };
+};
