+++ packages/vscode/src/utils/csm.mapper.ts
@@ -3,8 +3,8 @@ import type {
   CSMDictionary,
   DifferenceStatistic,
 } from "@githru-vscode-ext/analysis-engine/src/types";
-
-import type { ClusterNode, CommitNode, DiffStatistics } from "./NodeTypes.temps";
+import type { DiffStatistics } from "../types/DiffStatistics";
+import type { ClusterNode, CommitNode } from "../types/Node";
 
 /**
  * engine DifferenceStatistic → view DiffStatistics
@@ -34,7 +34,7 @@ const mapDiffStatisticsFrom = (params: { differenceStatistic: DifferenceStatisti
 const mapCommitNodeListFrom = (params: { commits: StemCommitNode[]; clusterId: number }): CommitNode[] => {
   const { commits, clusterId } = params;
   return commits.map(({ commit }) => {
-    const releaseTags=commit.tags.filter(tag=>tag.startsWith('v')||/^[0-9.]+$/.test(tag));
+    const releaseTags = commit.tags.filter((tag) => tag.startsWith("v") || /^[0-9.]+$/.test(tag));
     return {
       nodeTypeName: "COMMIT" as const,
       commit: {
@@ -63,7 +63,7 @@ const mapCommitNodeListFrom = (params: { commits: StemCommitNode[]; clusterId: n
       // hasMajorTag: false,
       // hasMinorTag: false,
       clusterId,
-    }
+    };
   });
 };
